requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
supabase>=2.15.0
anthropic>=0.43.0
python-dotenv>=1.0.1
